Extracted IdentityView from IdentityViewController
